+++ core/providers/gigachat/gigachat_test.go
@@ -29,6 +29,7 @@ func TestGigachat(t *testing.T) {
 	t.Run("Responses", testGigaChatResponses)
 	t.Run("ResponsesStream", testGigaChatResponsesStream)
 	t.Run("Tools", testGigaChatTools)
+	t.Run("Errors", testGigaChatErrors)
 	t.Run("BuildsTLSClientWithCABundle", testGigaChatBuildsTLSClientWithCABundle)
 	t.Run("BuildsTLSClientWithCertificate", testGigaChatBuildsTLSClientWithCertificate)
 	t.Run("RejectsMissingCertificatePair", testGigaChatRejectsMissingCertificatePair)